recursive_root/requirements.txt
openai
requests
anytree
pandas
tiktoken
numpy
qdrant-client
python-dotenv
